ユーザーの解答、問題・解答検索一覧のajax通信中に表示するspinnerの位置を変更
