+++ app/javascript/src/questions/Index.js
@@ -45,23 +45,4 @@ $(function(){
     var dropDownButton = $('.search-form-universities button');
     displayCheckedUniversityNamesOnDropDownButton(checkedBoxes, dropDownButton);
   });
-
-  const scrollToElem = document.querySelector('.scroll-to');
-  const spinner = $('<div class="spinner-border d-inline ms-4", style="width: 1.5rem; height: 1.5rem;", role="status">');
-  // ajax開始前
-  document.body.addEventListener("ajax:before", (event) => {
-    if(event.target.classList.contains("loading")){
-      // 問題一覧でのスピナー
-      $(scrollToElem).append(spinner);
-      scrollToElem.scrollIntoView();
-      // ajax終了時
-      event.target.addEventListener("ajax:success", (event) => {
-        // 問題一覧でのスピナー
-        const spinner = $('.scroll-to .spinner-border');
-        if(spinner){
-          spinner.remove();
-        }
-      });
-    }
-  });
 });
