+++ app/views/answers/_answers.html.slim
@@ -1,6 +1,6 @@
 / 画面サイズmd以上：2列、md以下:1列
 .container
-  .row id="answers-index"
+  .row.loaded id="answers-index"
     / カード部分
     = content_for :answer_partial
