+++ app/views/shared/questions/_index.html.slim
@@ -97,7 +97,7 @@
 
   / 問題カードの一覧
   .container
-    .row. id="questions-index"
+    .row.loaded id="questions-index"
       = content_for :question_partial
 
 / ページネーション
